Move function into auth folder
